+++ UI/Web/src/app/collections/all-collections/all-collections.component.ts
@@ -3,13 +3,13 @@ import { Title } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { EditCollectionTagsComponent } from 'src/app/cards/_modals/edit-collection-tags/edit-collection-tags.component';
-import { UtilityService } from 'src/app/shared/_services/utility.service';
 import { CollectionTag } from 'src/app/_models/collection-tag';
 import { JumpKey } from 'src/app/_models/jumpbar/jump-key';
 import { Tag } from 'src/app/_models/tag';
 import { ActionItem, ActionFactoryService, Action } from 'src/app/_services/action-factory.service';
 import { CollectionTagService } from 'src/app/_services/collection-tag.service';
 import { ImageService } from 'src/app/_services/image.service';
+import { JumpbarService } from 'src/app/_services/jumpbar.service';
 
 
 @Component({
@@ -30,7 +30,7 @@ export class AllCollectionsComponent implements OnInit {
 
   constructor(private collectionService: CollectionTagService, private router: Router,
     private actionFactoryService: ActionFactoryService, private modalService: NgbModal, 
-    private titleService: Title, private utilityService: UtilityService, 
+    private titleService: Title, private jumpbarService: JumpbarService, 
     private readonly cdRef: ChangeDetectorRef, public imageSerivce: ImageService) {
     this.router.routeReuseStrategy.shouldReuseRoute = () => false;
     this.titleService.setTitle('Kavita - Collections');
@@ -54,7 +54,7 @@ export class AllCollectionsComponent implements OnInit {
     this.collectionService.allTags().subscribe(tags => {
       this.collections = tags;
       this.isLoading = false;
-      this.jumpbarKeys = this.utilityService.getJumpKeys(tags, (t: Tag) => t.title);
+      this.jumpbarKeys = this.jumpbarService.getJumpKeys(tags, (t: Tag) => t.title);
       this.cdRef.markForCheck();
     });
   }
